split moonbase into tabs
