+++ packages/core-extensions/src/moonbase/ui/config/index.tsx
@@ -1,0 +1,9 @@
+import { WebpackRequireType } from "types/src";
+
+export default (require: WebpackRequireType) => {
+  const React = require("common_react");
+
+  return function ConfigPage() {
+    return <>config soon:tm:</>;
+  };
+};
